Fixes, reverted back to non-parallel hash table verification to prevent memory leaks

+++ objdetect/matcher.h
@@ -31,10 +31,10 @@ namespace tless {
         void nonMaximaSuppression(std::vector<Match> &matches);
 
         // Tests
-        int testObjectSize(float scale, float depth, Window &window, cv::Mat &sceneDepth, cv::Point &stable); // Test I
+        int testObjectSize(float scale, ushort depth, Window &window, cv::Mat &sceneDepth, cv::Point &stable); // Test I
         int testSurfaceNormal(uchar normal, Window &window, cv::Mat &sceneSurfaceNormalsQuantized, cv::Point &stable); // Test II
         int testGradients(uchar gradient, Window &window, cv::Mat &sceneAnglesQuantized, cv::Mat &sceneMagnitudes, cv::Point &edge); // Test III
-        int testDepth(float scale, float diameter, float depthMedian, Window &window, cv::Mat &sceneDepth, cv::Point &stable); // Test IV
+        int testDepth(float scale, float diameter, ushort depthMedian, Window &window, cv::Mat &sceneDepth, cv::Point &stable); // Test IV
         int testColor(cv::Vec3b HSV, Window &window, cv::Mat &sceneHSV, cv::Point &stable); // Test V
     public:
         // Constructor
